requirements.txt
pillow == 9.40